requirements.txt
Markdown
matplotlib
numpy
pandas
psycopg2
scikit-learn
scipy
seaborn
SQLAlchemy
mock
pyyaml
